Day9

+++ include/types.hpp
@@ -1,6 +1,6 @@
 #include <cstdint>
 
-using i32 = std::uint64_t;
-using i64 = std::uint64_t;
-using u32 = std::uint64_t;
+using i32 = std::int32_t;
+using i64 = std::int64_t;
+using u32 = std::uint32_t;
 using u64 = std::uint64_t;
